Implement lazy converter map preparation.
So far, inside the participant configuration, the converter map was prepared
once initially with the first access to the participant config and afterwards
loaded from a cached variable. As a consequence it was impossible to add a
converter to the global converter map without regenerating the whole
ParticipantConfig. This, however, had the drawback that previous manual
modifications to the configuration were erased. With this change the converter
map for a new participant is created in a lazy fashion each time a converter
map for a participant is requested. Hence, new converters can now be added to
the global converter map and are automatically available for any new
participant.

* rsb/__init__.py: Implemented lazy loading of converter map
* test/coretest.py: Unit test adde to verify the behavior

refs #1408