Enable introspection without host id
Ensure that introspection can be used on platforms, where a unique host
id cannot be determined. This includes using the host name as a
surrogate in these cases and stripping of domain parts from the host
name, since a dot character cannot be used in introspection scopes.

Remove the assumption that a host id always exists from the unit tests.

Signed-off-by: Jan Moringen <jmoringe@techfak.uni-bielefeld.de>