Use TransportFactories for creating Connectors
Adds a new class TransportFactory which has the duty of creating
connector instances from given configuration options for a individual
transports. A factory instance for each transport is registered in the
rsb.transport module and the participant creation was changed to use
these factories. Transport modules now have an initialize method which
should register these factory instances.

Tests had to be adapted so that required transports are available.

fixes #1025