Convert unit tests to nose
From now on python setup.py nosetests needs to be used to run the unit
tests. This will automatically generate coverage and xunit-compliant
data. The output locations for these files have changed to stick to the
defaults of nose.

Spread is now required for testing as it is actually a bad idea to not
test everything and then get the impression that everything is working
correctly, despite having introduced an error in untested code.

Transport options for unit tests (e.g. ports or the spread daemon
executable) can now be specified using the nose-testconfig plugin.
Please refer to https://pypi.python.org/pypi/nose-testconfig for
instructions. The configuration keys are spread.port, spread.daemon,
socket.port.