Fix config parsing for unit tests
Due to the fact that nose executes modules once to discover available
tests, it is not safe to assume that at global module initialization
time things performed by the fixtures of the package have already been
done. In this case, the config files for transports had not been written
and caused the tests to fail.