Better message discrimination in rsb/transport/rsbspread/__init__.py
Use the presence of the msg_type attribute to identify "regular"
Spread messages. This has the advantage that it can be emulated by the
mock DummyMessage which is used in tests. Eventually, this will allow
simplifying the termination of the receiver loop.

In anticipation of this better termination behavior, no longer turn
the receiver thread into a daemon thread.

* rsb/transport/rsbspread/__init__.py (SpreadReceiverTask.__call__):
  use hasattr 'msg_type' to identity "regular" Spread messages
  (InPushConnector.activate): do not turn the receiver thread into a
  daemon thread
* test/rsbspread_test.py (SpreadConnectorTest.DummyMessage.__init__):
  new method; add a msg_type attribute