Integrate codecov for coverage reporting
Adds a coverage command to the tox test execution. The default command
just prints the python version as a dummy. The travis execution
overwrites this command with the call to the codecov reporting tool.

This commit assumes a suitable configuration of codecov (project is
registered) and travis (environment variable with the secret upload key
is created).