Replace double underscores with a single one
With PEP8 two underscores are only intended to avoid conflicts with
parent classes. Otherwise, a single underscore should be used to mark
variables as private. All instances have been converted.

This commit resolves one name clash in the socket server implementation
by exposing the active flag in Bus and reusing it in BusServer instead
of managing a custom flag with the same name and functionality.